LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/document-write/script-src-module.tentative.sub-expected.txt

CONSOLE MESSAGE: TypeError: 'text/plain' is not a valid JavaScript MIME type.

    <!-- speculative case in document.write -->
    <script src="/html/syntax/speculative-parsing/resources/stash.py?action=put&amp;uuid=7ab8729d-c0fd-498c-b9ed-ff3472434c50&amp;encodingcheck=&Gbreve;" type=module></script>



PASS Speculative parsing, document.write(): script-src-module
